pyproject.toml
[project]
name = "notes-improver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "markdown-full-yaml-metadata>=2.2.1",
    "openai>=1.107.3",
    "peek-python>=25.0.24",
    "python-frontmatter>=1.1.0",
    "python-markdown>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
